fix(coding-agent): abort native web search discovery
